Person.ini文件默认是会在c盘下面生成, 默认是会不断存储聊天的群众, 后续会加一个如果一天内, 好感值还是不变的话就删除这个记录 另外Person.ini是自动读取的了, 明天开始做个可以加好感的事件. 下一步打算: 1. 好感事件增加, 改变文件的接口完成; 2. 封装关于聊天的随机回应接口.

+++ CQPdemo/RrConfig.h
@@ -15,7 +15,7 @@ namespace rr
 		bool ReadConfig(const std::string& filename);
 		std::string ReadString(const char* section, const char* item, const char* default_value);
 		int ReadInt(const char* section, const char* item, const int& default_value);
-		float ReadFloat(const char* section, const char* item, const float& default_value);
+		double ReadFloat(const char* section, const char* item, const float& default_value);
 	private:
 		bool IsSpace(char c);
 		bool IsCommentChar(char c);
